requirements.txt
fastapi
langchain
langchain-google-genai
langserve
pydantic
uvicorn
python-dotenv 
protobuf==3.20.*  